ppc: Adjust isa_io_base for ppc64
To avoid cast warnings, use a 64-bit ISA I/O base on ppc64.

v3:
* Adjust unix target, pointed out by Blue.
  Unify ppc and sparc64, using unsigned long instead of uintptr_t.
  It is initialized from pci_arch_t, which uses unsigned long.
* Update other ppc targets, pointed out by Blue.

v2:
Initial.

Cc: Blue Swirl <blauwirbel@gmail.com>
Cc: Alexander Graf <agraf@suse.de>


git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@964 f158a5a8-5612-0410-a976-696ce0be7e32

+++ arch/ppc/briq/init.c
@@ -53,7 +53,7 @@ unexpected_excep( int vector )
 		;
 }
 
-uint32_t isa_io_base;
+unsigned long isa_io_base;
 
 void
 entry( void )
